[objc-appscript] improved error descriptions; AEMApplication no longer tries to clean up open transactions automatically on dealloc/finalize (clients must now ensure transactions are correctly ended in their own code); fixed return type for -hash methods
git-svn-id: https://appscript.svn.sourceforge.net/svnroot/appscript@561 5e899d65-cff2-4c64-b955-3fa930e4d869

+++ objc-appscript/trunk/src/Appscript/base.m
@@ -27,7 +27,7 @@
 }
 
 
-- (unsigned)hash {
+- (unsigned long)hash {
 	if (!cachedHash)
 		cachedHash = [[self description] hash];
 	return cachedHash;
